+++ cdrs-tokio/src/cluster/rustls_connection_manager.rs
@@ -1,4 +1,4 @@
-use async_trait::async_trait;
+use futures::FutureExt;
 use std::net::SocketAddr;
 use std::ops::Deref;
 use std::sync::Arc;
@@ -13,6 +13,7 @@ use crate::cluster::{KeyspaceHolder, NodeRustlsConfig};
 use crate::compression::Compression;
 use crate::error::Result;
 use crate::frame::Frame;
+use crate::future::BoxFuture;
 use crate::transport::{CdrsTransport, TransportRustls};
 
 pub struct RustlsConnectionManager {
@@ -25,48 +26,51 @@ pub struct RustlsConnectionManager {
     connection: RwLock<Option<Arc<TransportRustls>>>,
 }
 
-#[async_trait]
 impl ConnectionManager<TransportRustls> for RustlsConnectionManager {
-    async fn connection(
-        &self,
-        reconnection_policy: &ThreadSafeReconnectionPolicy,
-    ) -> Result<Arc<TransportRustls>> {
-        {
-            let connection = self.connection.read().await;
+    fn connection<'a>(
+        &'a self,
+        reconnection_policy: &'a ThreadSafeReconnectionPolicy,
+    ) -> BoxFuture<Result<Arc<TransportRustls>>> {
+        async move {
+            {
+                let connection = self.connection.read().await;
+                if let Some(connection) = connection.deref() {
+                    if !connection.is_broken() {
+                        return Ok(connection.clone());
+                    }
+                }
+            }
+
+            let mut connection = self.connection.write().await;
             if let Some(connection) = connection.deref() {
                 if !connection.is_broken() {
+                    // somebody established connection in the meantime
                     return Ok(connection.clone());
                 }
             }
-        }
 
-        let mut connection = self.connection.write().await;
-        if let Some(connection) = connection.deref() {
-            if !connection.is_broken() {
-                // somebody established connection in the meantime
-                return Ok(connection.clone());
-            }
-        }
+            let mut schedule = reconnection_policy.new_node_schedule();
 
-        let mut schedule = reconnection_policy.new_node_schedule();
+            loop {
+                let transport = self.establish_connection().await;
+                match transport {
+                    Ok(transport) => {
+                        let transport = Arc::new(transport);
+                        *connection = Some(transport.clone());
 
-        loop {
-            let transport = self.establish_connection().await;
-            match transport {
-                Ok(transport) => {
-                    let transport = Arc::new(transport);
-                    *connection = Some(transport.clone());
-
-                    return Ok(transport);
-                }
-                Err(error) => {
-                    let delay = schedule.next_delay().ok_or(error)?;
-                    sleep(delay).await;
+                        return Ok(transport);
+                    }
+                    Err(error) => {
+                        let delay = schedule.next_delay().ok_or(error)?;
+                        sleep(delay).await;
+                    }
                 }
             }
         }
+        .boxed()
     }
 
+    #[inline]
     fn addr(&self) -> SocketAddr {
         self.config.addr
     }